requirements.txt

torch
openai
tiktoken
python-dotenv
huggingface_hub
